+++ packages/fluent-theme/src/components/theme/Theme.module.css
@@ -609,7 +609,7 @@
     outline: var(--webchat-strokeWidthThick) solid var(--webchat-colorStrokeFocus2);
   }
 
-  &[aria-disabled='true'] {
+  :not(.webchat__thumb-button--is-submitted)[aria-disabled='true'] {
     color: var(--webchat-colorNeutralForegroundDisabled);
   }
 }
